+++ src/view/day_list.js
@@ -1,0 +1,7 @@
+const makeDayList = () => {
+  let dayList = `
+  <ul class="trip-days">
+  </ul>`;
+  return dayList;
+};
+export {makeDayList};
