+++ src/view/navigation.js
@@ -1,0 +1,9 @@
+const makeNavigation = () => {
+  let navigation = `
+  <nav class="trip-controls__trip-tabs  trip-tabs">
+    <a class="trip-tabs__btn  trip-tabs__btn--active" href="#">Table</a>
+    <a class="trip-tabs__btn" href="#">Stats</a>
+  </nav>`;
+  return navigation;
+};
+export {makeNavigation};
